Split class to own file
