+++ src/SearchEngine.Lucene.Bootstrap/QueryInfo.cs
@@ -1,0 +1,31 @@
+﻿namespace SearchEngine.Lucene.Bootstrap
+{
+    using System;
+    using System.Collections.Generic;
+    using System.Diagnostics;
+    using System.Linq;
+
+    using SearchEngine.Interface;
+
+    [DebuggerDisplay("{QueryType.Name,nq}")]
+    public sealed class QueryInfo
+    {
+        public readonly Type QueryType;
+
+        public readonly Type ResultType;
+
+        public QueryInfo(Type queryType)
+        {
+            QueryType = queryType;
+            ResultType = DetermineResultTypes(queryType).Single();
+        }
+
+        public static bool IsQuery(Type type) => DetermineResultTypes(type).Any();
+
+        private static IEnumerable<Type> DetermineResultTypes(Type type) =>
+            from interfaceType in type.GetInterfaces()
+            where interfaceType.IsGenericType
+            where interfaceType.GetGenericTypeDefinition() == typeof(ISearchEngineQuery<>)
+            select interfaceType.GetGenericArguments()[0];
+    }
+}
